+++ CHANGELOG.md
@@ -6,6 +6,7 @@ file.
 ## [Unreleased]
 ### Added
 - Now trace into executed binaries
+- Added `--avoid-cfg-tarpaulin` flag to remove `--cfg=tarpaulin` from the `RUSTFLAGS`
 
 ### Changed
 - Address offset mapping has been added which allows us to compile binaries